# docs/swagger.yaml
definitions:
  models.User:
    properties:
      created_by:
        type: string
      created_on:
        type: integer
      deleted_at:
        type: integer
      id:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      name:
        type: string
    type: object
  models.UserSwagger:
    properties:
      lists:
        items:
          $ref: '#/definitions/models.User'
        type: array
      total:
        type: integer
    type: object
  tool.JSONRET:
    properties:
      data:
        type: object
      error_code:
        type: integer
      msg:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: 18211167516@163.com
    name: baichonghua
  description: gin-web框架
  license: {}
  termsOfService: https://github.com/18211167516/Go-Gin-Api
  title: go-api 框架
  version: "1.0"
paths:
  /api/v1/user/{id}:
    get:
      description: 获取单个用户
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: 查询成功
          schema:
            $ref: '#/definitions/models.UserSwagger'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20001":
          description: Token鉴权失败
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20002":
          description: Token已超时
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20004":
          description: Token错误
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20005":
          description: Token参数不能为空
          schema:
            $ref: '#/definitions/tool.JSONRET'
      summary: 单个用户
      tags:
      - 用户信息
  /api/v1/users:
    get:
      description: 获取用户列表
      produces:
      - application/json
      responses:
        "0":
          description: 查询成功
          schema:
            $ref: '#/definitions/models.UserSwagger'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20001":
          description: Token鉴权失败
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20002":
          description: Token已超时
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20004":
          description: Token错误
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20005":
          description: Token参数不能为空
          schema:
            $ref: '#/definitions/tool.JSONRET'
      summary: 用户列表
      tags:
      - 用户信息
    post:
      consumes:
      - multipart/form-data*
      description: 新增用户
      parameters:
      - description: Name
        in: formData
        name: name
        required: true
        type: string
      - description: CreatedBy
        in: formData
        name: created_by
        required: true
        type: string
      responses:
        "0":
          description: 创建成功
          schema:
            $ref: '#/definitions/models.UserSwagger'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20001":
          description: Token鉴权失败
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20002":
          description: Token已超时
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20004":
          description: Token错误
          schema:
            $ref: '#/definitions/tool.JSONRET'
        "20005":
          description: Token参数不能为空
          schema:
            $ref: '#/definitions/tool.JSONRET'
      summary: 新增用户
      tags:
      - 用户信息
swagger: "2.0"